Criando classe Combat
